Refactor Proxies
