+++ app/src/main/java/ayds/newyork/songinfo/moredetails/data/external/proxies/Proxy.kt
@@ -1,4 +1,4 @@
-package ayds.newyork.songinfo.moredetails.data.external
+package ayds.newyork.songinfo.moredetails.data.external.proxies
 
 import ayds.newyork.songinfo.moredetails.domain.entities.Card
